move onClose logic into Dialog, reduce Location.js
